requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

llama_cpp_python==0.3.16
psutil==7.0.0
